pyproject.toml
[project]
name = "airflow-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "apache-airflow>=2.10.4",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
]
